requirements.txt
docx2txt
ipykernel
numpy==1.21.2
pandas==1.3.3
python-docx
